ffmpeg-worker/requirements.txt
# FFmpeg Worker Dependencies
flask==3.0.0
gunicorn==21.2.0

# Google Cloud
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-cloud-speech==2.21.0
google-auth==2.23.4

# HTTP client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
